making example consistent with other adapters
