+++ example/app.js
@@ -1,0 +1,28 @@
+var feathers = require('feathers');
+var bodyParser = require('body-parser');
+var mongoose = require('../lib');
+var Todo = require('./models/todo');
+
+// Create a feathers instance.
+var app = feathers()
+  // Enable Socket.io
+  .configure(feathers.socketio())
+  // Enable REST services
+  .configure(feathers.rest())
+  // Turn on JSON parser for REST services
+  .use(bodyParser.json())
+  // Turn on URL-encoded parser for REST services
+  .use(bodyParser.urlencoded({extended: true}))
+
+// Connect to the db, create and register a Feathers service.
+app.use('todos', mongoose('todo', Todo));
+
+// A basic error handler, just like Express
+app.use(function(error, req, res, next){
+  res.json(error);
+});
+
+// Start the server
+module.exports = app.listen(3030);
+
+console.log('Feathers Todo mongoose service running on 127.0.0.1:3030');
