+++ example/models/todo.js
@@ -1,0 +1,21 @@
+var mongoose = require('mongoose');
+
+var Todo = {
+    schema: {
+        text: {type: String, required: true},
+        complete: {type: Boolean, 'default': false, index: true},
+        createdAt: {type: Date, 'default': Date.now},
+        updatedAt: {type: Date, 'default': Date.now}
+    },
+    methods: {
+    },
+    statics: {
+    },
+    virtuals: {
+    },
+    indexes: [
+        {'updatedAt': -1, background: true}
+    ]
+};
+
+module.exports = Todo;
